requirements.txt
selenium
chromedriver_binary
python-dotenv
beautifulsoup4